[fix] Handle LOCAL INFILE read errors
Adds an error event listener to the LOCAL INFILE read stream to handle read
errors generated midway through reading the file. Also removes the previous
fs.access(...) check as there is a race condition between validating whether
the file exists and is readable and actually trying to read it.

If the file does not exist, is not readable, or some other error occurs while
trying to read it then they will all be handled the same way.