[CONJS-84] option `restoreNodeTimeout` is not respected when removeNodeErrorCount is set
Goal is to correct wrong behaviour when setting restoreNodeTimeout:
cluster isn't blacklisting node if the option "removeNodeErrorCount"
was set, only after number of error reach "removeNodeErrorCount" value.

Correction is to make that corresponds to documentation: when a node
fails, it will be blacklisted, so the connection will be done on other
nodes. If "removeNodeErrorCount" value is set and reached, the node will
 be removed.