[CONJS-138] pool.getConnection() might not timeout even with acquireTimeout set #116 Pool option `acquireTimeout` ensure that pool.getConnection() throw an error if there is no available connection after this timeout is reached.
When connection stays in pool unused for some time (`minDelayValidation` option), pool will return those connection after a validation (PING to server). The issue is that when using proxy between client and server, socket to proxy might still be open, but proxy might be busy reconnecting server for an unknown amount of time. In those case, `acquireTimeout` is not respected, and pool.getConnection() might wait indefinitely.

New option `pingTimeout`, defaulting to 250ms is added so validation is limited in time.

Connection.ping() now has an optional Timeout parameter. If timeout is reached, promise is rejected and connection is force closed.