[CONJS-133] Support ES2020 BigInt object
BigInt is a built-in object that provides a way to represent whole numbers larger than 2^53 - 1, which is the largest number JavaScript can reliably represent with the Number primitive and represented by the Number.MAX_SAFE_INTEGER constant. BigInt can be used for arbitrarily large integers.

There is 3 different parts in implementation:
* BigInt parameters are now supported.
* resultset that have data of MariaDB BIGINT type will be returned as BigInt type in place of Number.
* UPSERT query resulting insertId has BigInt format

This is only enabled when option `supportBigInt` is enable for compatibility with previous implementation. When enable, option `supportBigNumbers` has no used anymore.
This will be the default in next MAJOR 3.x version.