[CONJS-77] sha256Password and caching_sha256_password authentication plugin implementation
This permit MySQL authentication with 5.7 / 8.0

Node.js v11.6+ is needed in order to perform encryption. An error will be thrown in case Node is not compatible and use those authentication plugin.
if SSL is enable, no key is required. if not, public RSA key must be provided to client using options `cachingRsaPublicKey` or `rsaPublicKey.

example of use :
sql : CREATE USER 'cachingSha256User2'@'%' IDENTIFIED WITH caching_sha2_password BY 'password';

    const conn = await mariadb.createConnection({
        user: 'sha256User',
        password: 'password',
        rsaPublicKey: '/somepath/public.pem'
      });