[CONJS-196] 3.0 regression on release connection #195
In 3.0, when releasing connection to pool not with connection.release() but connection.close(), connection wasn't given back to pool, but killed.
When killing connection, server send a ER_CONNECTION_KILLED error that can sometime be caught by connector before socket is closed, but that's not assured, then having just an ER_SOCKET_UNEXPECTED_CLOSE error.